+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Add Object and Array structure validators
 * Remove data.ValidateInterface...; does not work as expected
 * Remove structure.Validating; does not work as expected
 * Refactor data normalizer to use common structure normalizer